Add journal title abbreviation feature
It is often desirable to abbreviate journal names. Some references
already come with abbreviated journal names, but not all. Moreover, it
is often difficult to guess the correct abbreviation, in particular in
cases of foreign journals.

The ISO4 standard defines a set of rules to generate title word
abbreviations. The journal title abbreviation feature of BibTexTools
generates approximations of the ISO4 standard while being as simple and
as fast as possible. For this, the correct abbreviation patterns are
"guessed" by concerning only closest matches in the list of title
word abbreviations and checking them for exact matches. Some other rules
of the ISO4 standard are applied:

* articles, conjunctions and prepositions are omitted (except at the
  beginning),
* commas are removed, other punctuation is preserved,
* diacritics are left as in the original title,
* words like Part, Series, Section should be omitted,
* single word titles (possibly with an article or preposition) are not
  abbreviated,
* abbreviated words are capitalized.