Improve journal title abbreviations
Title word abbreviations should be capitalized, however, only unless the
title word consists only of uppercase characters; in this case the title
word is most likely an acronym. Also the "arXiv e-prints" journal name
should not be capitalized. Hence, if the journal title contains
"arXiv" it is not abbreviated.

To improve the overall performance of the "BibTeXTools: Format" command,
it is assumed that title words ending with periods are already
abbreviated. Although punctuation should be preserved in general, there
should be only rare occurrences of periods after non-abbreviated title
words.