Add async commands and thread progress indicator
Running a command will block the UI until the command is finished. By
running the commands on a separate thread, no UI freezing should occur
in principle. Additionally the progress of the thread execution can be
indicated in the status bar.

However, the processing of the active view still blocks the UI, although
the threading is working properly. This should be changed in the
future.